+++ src/unpath.sh
@@ -2,15 +2,26 @@ helpMsg="\
 The unpath tool is to expand file paths into their contents.
 
 unpath
+    [--path-prefix <path_prefix>]
+    [--path-suffix <path_suffix>]
+    [--prefix <prefix>]
+    [--suffix <prefix>]
     [-d | --document-format <document_format>]
     [-h | --help]
     <root_directory_path>
     [<document_path>]
 
-Print a <document_path> file with each file path marker \
-appended with decorated contents of the file \
-read from a <root_directory_path> directory \
-by a <local_file_path> path.
+Print a <document_path> file
+with each <file_path_prefix><local_file_path><file_path_suffix> file path marker \
+prepended \
+with a <prefix> prefix\
+and appended \
+with contents \
+of the file, \
+read \
+from a <root_directory_path> directory \
+by a <local_file_path> path, \
+and a <suffix> suffix.
 
 PREDEFINED DOCUMENT FORMATS
 
@@ -20,11 +31,23 @@ Markdown
     file path markers                  -- <!-- \"<local_file_path>\" -->, \
 <!-- '<local_file_path>' -->
 
+--path-prefix (<\!--.*['\"])
+    a search pattern before a <local_file_path> path inside a file path marker
+
+--path-suffix (['\"].*-->)
+    a search pattern after a <local_file_path> path inside a file path marker
+
+--prefix ("'```'"<<root_directory_path>first_file_extension>)
+    text to prepend to inserted file contents
+
+--suffix ("'```'")
+    text to append to inserted file contents
+
 -d, --document-format (Markdown)
-    a predefined set of decorations to use
+    a predefined set of prefixes and suffixes to use
 
 -h, --help (0)
-    whether to print the help message and then exit\
+    whether to print the help message and then exit
 "
 noFlagOrOptErr () {
     echo "error: $1 flag or option undefined"
@@ -38,6 +61,10 @@ stdinDocMsg='document read from stdin...'
 args=()
 documentFormat=Markdown
 help=0
+pathPrefix=''
+pathSuffix=''
+prefix=''
+suffix=''
 
 while (( $# > 0 ))
 do
@@ -51,6 +78,26 @@ do
             help=1
             echo "$helpMsg"
             exit 0
+            ;;
+        --path-prefix)
+            pathPrefix=$2
+            shift
+            shift
+            ;;
+        --path-suffix)
+            pathSuffix=$2
+            shift
+            shift
+            ;;
+        --prefix)
+            prefix=$2
+            shift
+            shift
+            ;;
+        --suffix)
+            suffix=$2
+            shift
+            shift
             ;;
         -* | --*)
             echo "$(noFlagOrOptErr "$1")"
@@ -90,8 +137,14 @@ case $documentFormat in
         exit 1
         ;;
 esac
+: ${pathPrefix:="$pathPfx"}
+: ${pathSuffix:="$pathSfx"}
+: ${prefix:="$pfx"}
+: ${suffix:="$sfx"}
+
 fstMrkPath=$(echo "$doc" \
-            | sed --quiet --regexp-extended "s|$pathPfx(.*)$pathSfx|\1|p" \
+            | sed --quiet --regexp-extended \
+                  "s|$pathPrefix(.*)$pathSuffix|\1|p" \
             | head -1
             )
 fstLclRoot=${fstMrkPath%%/*}
@@ -100,13 +153,13 @@ for path in "${paths[@]}"
 do
     pathMrk=$pathPfx$path$pathSfx
     pathExpanded=$(echo "$pathExpanded" \
-                  | sed "\|$pathMrk|a $pfx\n$sfx" \
+                  | sed "\|$pathMrk|a $prefix\n$suffix" \
                   | sed "\|$pathMrk|N; \|\n|r $path"
                   )
     lclPath=$fstLclRoot${path##*/$fstLclRoot}
     lclPathMrk=$pathPfx$lclPath$pathSfx
     pathExpanded=$(echo "$pathExpanded" \
-                  | sed "\|$lclPathMrk|a $pfx\n$sfx" \
+                  | sed "\|$lclPathMrk|a $prefix\n$suffix" \
                   | sed "\|$lclPathMrk|N; \|\n|r $path"
                   )
 done